Remove the aggressive array preallocation. Improve benchmark result 100%

Before change
```
go test -bench=BenchmarkWatchableStorePut -run=xxx -benchmem -cpu=1,2,4
PASS
BenchmarkWatchableStorePut  	   50000	     28181 ns/op	   12157 B/op	      19 allocs/op
BenchmarkWatchableStorePut-2	  100000	     21586 ns/op	   12032 B/op	      17 allocs/op
BenchmarkWatchableStorePut-4	  100000	     19479 ns/op	   12028 B/op	      18 allocs/op
ok  	github.com/coreos/etcd/mvcc	9.143s
```

After change
```
go test -bench=BenchmarkWatchableStorePut -run=xxx -benchmem -cpu=1,2,4
PASS
BenchmarkWatchableStorePut  	  100000	     12463 ns/op	    2014 B/op	      19 allocs/op
BenchmarkWatchableStorePut-2	  200000	      9679 ns/op	    1836 B/op	      17 allocs/op
BenchmarkWatchableStorePut-4	  200000	      9222 ns/op	    1877 B/op	      18 allocs/op
ok  	github.com/coreos/etcd/mvcc	10.716s
```

@gyuho Probably you want to rerun the e2e benchmark to verify the change?