@gyuho 2016-04-30
@xiang90 LGTM. Below are the bench results:

```
# original
go test -bench=BenchmarkWatchableStorePut -run=xxx -benchmem -cpu=1,2,4
PASS
BenchmarkWatchableStorePut  	  100000	     24558 ns/op	   11857 B/op	      16 allocs/op
BenchmarkWatchableStorePut-2	  100000	     21278 ns/op	   11855 B/op	      16 allocs/op
BenchmarkWatchableStorePut-4	  100000	     16363 ns/op	   11878 B/op	      16 allocs/op

# with patch
go test -bench=BenchmarkWatchableStorePut -run=xxx -benchmem -cpu=1,2,4
PASS
BenchmarkWatchableStorePut  	  200000	     10196 ns/op	    1713 B/op	      16 allocs/op
BenchmarkWatchableStorePut-2	  200000	      9471 ns/op	    1710 B/op	      16 allocs/op
BenchmarkWatchableStorePut-4	  200000	      9417 ns/op	    1704 B/op	      16 allocs/op
```

![01-avg-cpu](https://cloud.githubusercontent.com/assets/6799218/14937417/5010a28c-0ebb-11e6-9aab-1d9396a13ac4.png)
![01-avg-latency-ms](https://cloud.githubusercontent.com/assets/6799218/14937419/5013d1b4-0ebb-11e6-9aea-1681e8c4e91e.png)
![01-avg-memory](https://cloud.githubusercontent.com/assets/6799218/14937420/5015dba8-0ebb-11e6-85b3-8c37d8d60e69.png)
![01-throughput](https://cloud.githubusercontent.com/assets/6799218/14937418/5011ba46-0ebb-11e6-9311-0fd7258bd1e7.png)